+++ NEWS.md
@@ -5,6 +5,16 @@
     -   `catr_atom_ad()`, `catr_atom_ad_db_all()`, `catr_atom_cp()`,
         `catr_atom_cp_db_all()`, `catr_atom_bu()`, `catr_atom_bu_db_all()`.
 
+-   Deprecate previous functions in favor of the new API:
+
+    -   New SRS database on `catr_srs_values`, replaces `coordinates`.
+
+    -   `catr_ovc_rccoor_distancia()` replaces `near_rc()`.
+
+    -   `catr_ovc_rccoor()` replaces `get_rc()`.
+
+-   Add precomputed vignettes.
+
 # CatastRo 0.1.0
 
 -   Initial release.